Moved other methods out into modules
