+++ resource_processor/resources/commands.py
@@ -1,0 +1,108 @@
+import asyncio
+import json
+import logging
+from resources.helpers import get_installation_id
+from shared.config import ProcessorConfig
+from shared.logging import shell_output_logger
+
+
+def azure_login_command(config: ProcessorConfig):
+    local_login = None
+    # If running locally with use_local_creds enabled, use signed-in identity for authenticating with Service Bus
+    if config.use_local_creds == "true":
+        local_login = f"az account set --subscription {config.arm_subscription_id}"
+
+    # If running locally with use_local_creds disabled, use the Service Principal credentials
+    else:
+        local_login = f"az login --service-principal --username {config.arm_client_id} --password {config.arm_client_secret} --tenant {config.arm_tenant_id}"
+
+    # Use the Managed Identity when in VMSS identity
+    vmss_login = f"az login --identity -u {config.vmss_msi_id}"
+    command = vmss_login if config.vmss_msi_id else local_login
+    return command
+
+
+def azure_acr_login_command(config: ProcessorConfig):
+    return f"az acr login --name {config.registry_server.replace('.azurecr.io','')}"
+
+
+async def build_porter_command(config: ProcessorConfig, logger, msg_body, custom_action=False):
+    porter_parameter_keys = await get_porter_parameter_keys(msg_body)
+    porter_parameters = ""
+
+    if porter_parameter_keys is None:
+        logger.warning("Unknown porter parameters - explain probably failed.")
+    else:
+        for parameter_name in porter_parameter_keys:
+            # try to find the param in order of priorities:
+            # 1. msg parameters collection
+            # 2. config (e.g. terraform state env vars)
+            # 3. msg body root (e.g. id of the resource)
+            parameter_value = msg_body['parameters'].get(
+                parameter_name,
+                config[parameter_name] if hasattr(config, parameter_name) else msg_body.get(parameter_name))
+
+            # if still not found, might be a special case
+            # (we give a chance to the method above to allow override of the special handeling done below)
+            if parameter_value is None:
+                parameter_value = get_special_porter_param_value(parameter_name, msg_body)
+
+            # only append if we have a value, porter will complain anyway about missing parameters
+            if parameter_value is not None:
+                porter_parameters = porter_parameters + f" --param {parameter_name}=\"{parameter_value}\""
+
+    installation_id = get_installation_id(msg_body)
+
+    command_line = [f"{azure_login_command(config)} && {azure_acr_login_command(config)} && porter "
+                    # If a custom action (i.e. not install, uninstall, upgrade) we need to use 'invoke'
+                    f"{'invoke --action ' if custom_action else ''}"
+                    f"{msg_body['action']} \"{installation_id}\" "
+                    f" --reference {config.registry_server}/{msg_body['name']}:v{msg_body['version']}"
+                    f" {porter_parameters} --cred ./vmss_porter/azure.json --allow-docker-host-access --force"
+                    f" && porter show {installation_id}"]
+    return command_line
+
+
+async def build_porter_command_for_outputs(msg_body):
+    installation_id = get_installation_id(msg_body)
+    # we only need "real" outputs and use jq to remove the logs which are big
+    command_line = [f"porter show {installation_id} --output json | jq -c '. | select(.Outputs!=null) | .Outputs | del (.[] | select(.Name==\"io.cnab.outputs.invocationImageLogs\"))'"]
+    return command_line
+
+
+async def get_porter_parameter_keys(config: ProcessorConfig, logger, msg_body):
+    command = [f"{azure_login_command(config)} >/dev/null && \
+        {azure_acr_login_command(config)} >/dev/null && \
+        porter explain --reference {config.registry_server}/{msg_body['name']}:v{msg_body['version']} -ojson"]
+
+    proc = await asyncio.create_subprocess_shell(
+        ''.join(command),
+        stdout=asyncio.subprocess.PIPE,
+        stderr=asyncio.subprocess.PIPE,
+        env=config.porter_env)
+
+    stdout, stderr = await proc.communicate()
+    logging.info(f'get_porter_parameter_keys exited with {proc.returncode}]')
+    result_stdout = None
+    result_stderr = None
+
+    if stdout:
+        result_stdout = stdout.decode()
+        porter_explain_parameters = json.loads(result_stdout)["parameters"]
+        porter_parameter_keys = [item["name"] for item in porter_explain_parameters]
+        return porter_parameter_keys
+    if stderr:
+        result_stderr = stderr.decode()
+        shell_output_logger(result_stderr, '[stderr]', logger, logging.WARN)
+
+
+def get_special_porter_param_value(config: ProcessorConfig, parameter_name: str, msg_body):
+    # some parameters might not have identical names and this comes to handle that
+    if parameter_name == "mgmt_acr_name":
+        return config.registry_server.replace('.azurecr.io', '')
+    if parameter_name == "mgmt_resource_group_name":
+        return config.tfstate_resource_group_name
+    if parameter_name == "workspace_id":
+        return msg_body.get("workspaceId")  # not included in all messages
+    if parameter_name == "parent_service_id":
+        return msg_body.get("parentWorkspaceServiceId")  # not included in all messages
