+++ resource_processor/resources/helpers.py
@@ -1,0 +1,5 @@
+def get_installation_id(msg_body):
+    """
+    This is used to identify each bundle install within the porter state store.
+    """
+    return msg_body['id']
